requirements.txt
streamlit
pymongo
langchain
langchain-community
langchain-google-genai
langchain-mongodb
langchain-classic
pypdf
docx2txt
python-dotenv
cloudinary
keras<3.0.0